refactor(component): split audit.rs god-file into modules
